+++ NEWS.md
@@ -1,5 +1,6 @@
 # tabula 0.9.0.9000
 
+* ADD: Brainerd-Robinson coefficient of similarity
 * ADD: Zuni ceramics dataset
 * ADD: Mississippi ceramics dataset
 * ADD: BootCA S4 class to store partial bootstrap CA results